展示订单

+++ haohuo-service/haohuo-service-order/src/main/java/com/ebanma/cloud/order/model/AccountInfo.java
@@ -12,7 +12,7 @@ public class AccountInfo {
 
     private BigDecimal freight;
 
-    private Integer totalIntegral;
+    private long totalIntegral;
 
     private Integer moneyDeduction;
